api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["sorenkoelbaek <soren.iversen@lego.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
dynaconf = "^3.2.6"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pg8000 = "^1.31.2"
sqlmodel = "^0.0.22"
pydantic = {extras = ["email"], version = "^2.10.6"}
pytest = "^8.3.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.19"
bcrypt = "^4.2.1"
spotipy = "^2.25.1"
websockets = "^15.0.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"


[tool.poetry.group.dev.dependencies]
alembic = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
